pyproject.toml
[project]
name = "go4hive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bleach>=6.3.0",
    "django>=6.0.4",
    "hive-nectar>=0.2.15",
]
